% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piqp.R
\name{status_description}
\alias{status_description}
\title{Return the solver status description string}
\usage{
status_description(code)
}
\arguments{
\item{code}{a valid solver return code}
}
\value{
a status description string
}
\description{
Return the solver status description string
}
\examples{
status_description(1) ## for solved problem
status_description(-1) ## for max iterations limit reached
}
